+++ ta4j-core/src/test/java/org/ta4j/core/indicators/statistics/StandardDeviationIndicatorTest.java
@@ -50,4 +50,21 @@ public class StandardDeviationIndicatorTest extends AbstractIndicatorTest<Indica
         assertNumEquals(0, sdv.getValue(3));
         assertNumEquals(0, sdv.getValue(8));
     }
+
+    @Test
+    public void sampleStandardDeviationCanBeRequestedExplicitly() {
+        var sdv = StandardDeviationIndicator.ofSample(new ClosePriceIndicator(data), 4);
+
+        assertNumEquals(0, sdv.getValue(0));
+        assertNumEquals(Math.sqrt(0.5), sdv.getValue(1));
+        assertNumEquals(Math.sqrt(1.0), sdv.getValue(2));
+        assertNumEquals(Math.sqrt(1.66666666666667), sdv.getValue(3));
+        assertNumEquals(Math.sqrt(0.66666666666667), sdv.getValue(4));
+        assertNumEquals(Math.sqrt(0.33333333333333), sdv.getValue(5));
+        assertNumEquals(Math.sqrt(0.66666666666667), sdv.getValue(6));
+        assertNumEquals(Math.sqrt(0.66666666666667), sdv.getValue(7));
+        assertNumEquals(Math.sqrt(0.66666666666667), sdv.getValue(8));
+        assertNumEquals(Math.sqrt(4.66666666666667), sdv.getValue(9));
+        assertNumEquals(Math.sqrt(14), sdv.getValue(10));
+    }
 }